# specs/001-package-catalog/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Elsa Platform API
  version: 0.1.0
  description: Public and admin REST contracts for package, feature, compatibility, sync, validation, and approval workflows.
servers:
  - url: http://localhost:5000
security: []
...
paths:
  /api/packages:
    get:
      tags: [Public Packages]
      summary: List public packages
      parameters:
        - $ref: '#/components/parameters/SourceIdsQuery'
        - $ref: '#/components/parameters/PackageIdQuery'
        - $ref: '#/components/parameters/TagQuery'
        - $ref: '#/components/parameters/FeatureIdQuery'
        - $ref: '#/components/parameters/ElsaVersionQuery'
      responses:
        '200':
          description: Public packages containing only listed, approved, valid versions.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PackageSummary'
...
  /api/sources/{sourceId}/packages/{packageId}:
    get:
      tags: [Public Packages]
      summary: Get source-qualified public package details
      parameters:
        - $ref: '#/components/parameters/SourceIdPath'
        - $ref: '#/components/parameters/PackageIdPath'
      responses:
        '200':
          description: Public package details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageDetails'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/admin/sync/packages/{packageId}:
    post:
      tags: [Admin Sync]
      security:
        - ApiKeyAuth: []
      summary: Trigger sync for one package ID
      parameters:
        - $ref: '#/components/parameters/PackageIdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncRequest'
      responses:
        '202':
          $ref: '#/components/responses/SyncAccepted'
...
  /api/admin/packages:
    get:
      tags: [Admin Packages]
      security:
        - ApiKeyAuth: []
      summary: List packages including admin states
      responses:
        '200':
          description: Admin packages.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminPackageSummary'
  /api/admin/packages/{packageId}:
    get:
      tags: [Admin Packages]
      security:
        - ApiKeyAuth: []
      summary: Get admin package details
      parameters:
        - $ref: '#/components/parameters/PackageIdPath'
      responses:
        '200':
          description: Admin package details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPackageDetails'
  /api/admin/packages/{packageId}/approve:
    post:
      tags: [Admin Packages]
      security:
        - ApiKeyAuth: []
      summary: Approve package
      parameters:
        - $ref: '#/components/parameters/PackageIdPath'
      requestBody:
        $ref: '#/components/requestBodies/ApprovalRequest'
      responses:
        '200':
          $ref: '#/components/responses/ApprovalUpdated'
  /api/admin/packages/{packageId}/reject:
    post:
      tags: [Admin Packages]
      security:
        - ApiKeyAuth: []
      summary: Reject package
      parameters:
        - $ref: '#/components/parameters/PackageIdPath'
      requestBody:
        $ref: '#/components/requestBodies/ApprovalRequest'
      responses:
        '200':
          $ref: '#/components/responses/ApprovalUpdated'
  /api/admin/packages/{packageId}/versions/{version}/approve:
    post:
      tags: [Admin Packages]
      security:
        - ApiKeyAuth: []
      summary: Approve package version
      parameters:
        - $ref: '#/components/parameters/PackageIdPath'
        - $ref: '#/components/parameters/VersionPath'
      requestBody:
        $ref: '#/components/requestBodies/ApprovalRequest'
      responses:
        '200':
          $ref: '#/components/responses/ApprovalUpdated'
  /api/admin/packages/{packageId}/versions/{version}/reject:
    post:
      tags: [Admin Packages]
      security:
        - ApiKeyAuth: []
      summary: Reject package version
      parameters:
        - $ref: '#/components/parameters/PackageIdPath'
        - $ref: '#/components/parameters/VersionPath'
      requestBody:
        $ref: '#/components/requestBodies/ApprovalRequest'
      responses:
        '200':
          $ref: '#/components/responses/ApprovalUpdated'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
...
    SourceIdPath:
      name: sourceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PackageIdPath:
      name: packageId
      in: path
      required: true
      schema:
        type: string
    VersionPath:
      name: version
      in: path
      required: true
      schema:
        type: string
...
    PackageIdQuery:
      name: packageId
      in: query
      schema:
        type: string
    TagQuery:
      name: tag
      in: query
      schema:
        type: string
    FeatureIdQuery:
      name: featureId
      in: query
      schema:
        type: string
    ElsaVersionQuery:
      name: elsaVersion
      in: query
      schema:
        type: string
    SourceIdsQuery:
      name: sourceIds
      in: query
      schema:
        type: array
        items:
          type: string
          format: uuid
      style: form
      explode: true
  requestBodies:
    ApprovalRequest:
      required: false
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApprovalRequest'
  responses:
    NotFound:
      description: Resource not found or not visible.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    SyncAccepted:
      description: Sync accepted.
      content:
        application/json:
          schema:
            type: object
            properties:
              syncRunId:
                type: string
              status:
                type: string
                enum: [Running]
    ApprovalUpdated:
      description: Approval state updated.
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                enum: [Pending, Approved, Rejected]
              updatedAt:
                type: string
                format: date-time
  schemas:
...
    PackageSummary:
      type: object
      required: [packageId, displayName, latestVersion]
      properties:
        packageId:
          type: string
        source:
          $ref: '#/components/schemas/PackageSourceProvenance'
        displayName:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        latestVersion:
          type: string
    PackageDetails:
      allOf:
        - $ref: '#/components/schemas/PackageSummary'
        - type: object
          properties:
            versions:
              type: array
              items:
                $ref: '#/components/schemas/PackageVersionSummary'
            documentation:
              $ref: '#/components/schemas/Documentation'
            license:
              $ref: '#/components/schemas/License'
    PackageVersionSummary:
      type: object
      required: [packageId, version]
      properties:
        packageId:
          type: string
        version:
          type: string
        source:
          $ref: '#/components/schemas/PackageSourceProvenance'
        publishedAt:
          type: string
          format: date-time
        indexedAt:
          type: string
          format: date-time
        featureCount:
          type: integer
...
    PackageSourceProvenance:
      type: object
      required: [id, name, url]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        url:
          type: string
...
    License:
      type: object
      properties:
        expression:
          type: string
        url:
          type: string
        requiresAcceptance:
          type: boolean
    Documentation:
      type: object
      properties:
        readmeUrl:
          type: string
        projectUrl:
          type: string
        releaseNotesUrl:
          type: string
        configurationUrl:
          type: string
...
    SyncRequest:
      type: object
      properties:
        forceReindex:
          type: boolean
          default: false
...
    AdminPackageSummary:
      type: object
      properties:
        packageId:
          type: string
        sourceId:
          type: string
        approved:
          type: boolean
        listed:
          type: boolean
        latestVersion:
          type: string
    AdminPackageDetails:
      allOf:
        - $ref: '#/components/schemas/AdminPackageSummary'
        - type: object
          properties:
            versions:
              type: array
              items:
                type: object
                properties:
                  version:
                    type: string
                  validationStatus:
                    type: string
                  approvalStatus:
                    type: string
                  isListed:
                    type: boolean
                  suspiciousChangeDetected:
                    type: boolean
    ApprovalRequest:
      type: object
      properties:
        reason:
          type: string
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string